Separate session from authz middleware
